refactor(jni): reorganize JNI structure and introduce new ABI module
- Restructured JNI modules for better clarity, moving support functions to a new `abi` module.
- Added panic and `Outcome` guards to enhance error handling across JNI boundaries.
- Implemented Java string marshalling utilities to streamline interactions with Java.
- Updated engine-related JNI exports to utilize the new ABI functions, improving code maintainability.
- Removed deprecated support module, consolidating functionality for a cleaner architecture.
