+++ crates/funput-jni/src/engine/lifecycle.rs
@@ -3,8 +3,8 @@
 use jni::EnvUnowned;
 use jni::sys::jlong;
 
-use crate::registry;
-use crate::support::{JavaObject, safe};
+use super::registry;
+use crate::abi::{JavaObject, safe};
 
 #[unsafe(no_mangle)]
 pub extern "system" fn Java_app_funput_funput_ime_nativebridge_FunputNative_nativeCreate(
