+++ crates/funput-jni/src/engine/mod.rs
@@ -1,0 +1,17 @@
+//! Composition JNI surface: the `FunputNative_native*` entry points and the
+//! ID-based ownership they run on.
+//!
+//! - [`registry`] — ID → `Engine` ownership (`create` / `destroy` / `with_mut`);
+//!   the JNI analogue of an opaque handle, so a stale or repeated destroy is a
+//!   safe no-op.
+//! - [`lifecycle`] — `nativeCreate` / `nativeDestroy` / `nativeClear`.
+//! - [`composition`] — `nativeProcess` / `nativeBoundary` / `nativeBackspace`.
+//! - [`settings`] — `nativeSetMethod` / `nativeSetToneStyle` / the boolean toggles.
+//!
+//! Exports link by `#[no_mangle]` symbol name, so their module location is
+//! invisible to the Android runtime.
+
+mod composition;
+mod lifecycle;
+mod registry;
+mod settings;
