+++ crates/funput-jni/src/engine/registry.rs
